css bugado

+++ back/src/main/java/br/com/back/controle/ProdutoControle.java
@@ -1,5 +1,8 @@
 package br.com.back.controle;
 
+
+import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -9,7 +12,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
@@ -29,6 +31,11 @@ public class ProdutoControle {
         return produtoDAO.findAll();
     }
 
+    private byte[] imagem(Integer id){
+        Optional<ProdutoModelo> produtoModelo = produtoDAO.findById(id);
+        return produtoModelo.orElse(null).getImg();
+    }
+
     @PostMapping
     public ResponseEntity<ProdutoModelo> cadastrar(@RequestParam("img") MultipartFile img, @RequestParam("descricao") String descricao, @RequestParam("preco") Double preco, @RequestParam("tipo") Integer tipo){
         try {
@@ -46,12 +53,35 @@ public class ProdutoControle {
     }
 
     @PutMapping
-    public ResponseEntity<ProdutoModelo> alterar(@RequestBody ProdutoModelo produtoModelo){
-        return new ResponseEntity<ProdutoModelo>(produtoDAO.save(produtoModelo), HttpStatus.ACCEPTED);
+    public ResponseEntity<ProdutoModelo> alterar(@RequestParam("img") MultipartFile img, @RequestParam("id") Integer id, @RequestParam("descricao") String descricao, @RequestParam("preco") Double preco, @RequestParam("tipo") Integer tipo){
+        try {
+            byte[] imgByte;
+            ProdutoModelo modelo = new ProdutoModelo();
+            if(img.isEmpty()){
+                imgByte = imagem(id);
+            }else{
+                imgByte = img.getBytes();
+            }
+            modelo.setImg(imgByte);
+            modelo.setDescricao(descricao);
+            modelo.setPreco(preco);
+            modelo.setTipo(tipo);
+            modelo.setId(id);
+            return new ResponseEntity<ProdutoModelo>(produtoDAO.save(modelo), HttpStatus.CREATED);
+        } catch (Exception e) {
+            // TODO: handle exception
+            return null;
+        }        
     }
     
     @DeleteMapping("/{id}")
-    public void remover(@PathVariable Integer id){
-        produtoDAO.deleteById(id);
+    public ResponseEntity<ProdutoModelo> remover(@PathVariable Integer id){
+        ProdutoModelo produto = produtoDAO.findById(id).orElse(null);
+        if(id != null){
+            produtoDAO.deleteById(id);
+            return new ResponseEntity<>(produto, HttpStatus.OK);
+        }else{
+            return new ResponseEntity<ProdutoModelo>(HttpStatus.NOT_FOUND);
+        }
     }
 }
